pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "merovingian"
version = "0.1.0"
description = "Cross-repository dependency intelligence for AI agents via MCP"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pyyaml>=6.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
merovingian = "merovingian.cli.app:main"
merovingian-mcp = "merovingian.mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/merovingian"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "PLW"]

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["merovingian"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["merovingian"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_also = ["if __name__", "def main"]
omit = [
    "src/merovingian/mcp/__main__.py",
    "src/merovingian/mcp/server.py",
]
